requirements.txt
pipecat-ai[webrtc,silero,openai,local-smart-turn-v3,runner]==0.0.100
pipecat-ai-cli
python-dotenv
fastmcp
httpx
torch
torchaudio
torchvision
motor
pydantic-settings
langchain
langchain-openai
sentence-transformers
pdfplumber
python-docx
psycopg2-binary
requests
beautifulsoup4
qdrant-client
transformers
langchain-text-splitters
langchain-qdrant
FlagEmbedding